pyproject.toml
[project]
name = "final_project"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "KyungMin.Ko", email = "rhrudals312@naver.com"},
    {name = "geza", email = ""},
]
dependencies = [
    "django>=5.1.1",
    "djangorestframework>=3.15.2",
    "pillow>=10.4.0",
    "pre-commit>=3.8.0",
    "django-environ>=0.11.2",
    "psycopg2>=2.9.9",
    "reportlab>=4.2.2",
    "requests>=2.32.3",
    "djangorestframework-simplejwt>=5.3.1",
    "boto3>=1.35.26",
    "django-cors-headers>=4.4.0",
    "psycopg2-binary>=2.9.9",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false


[tool.pdm.scripts]
m = "python src/manage.py"
r = "python src/manage.py runserver"
migs =  "python src/manage.py makemigrations"
mig =  "python src/manage.py migrate"
